pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "selenium>=4.38.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "reportlab>=4.0.0",
    "markdown>=3.5.0",
    "zhconv>=1.4.2",
    "yt-dlp>=2024.0.0",
    "whisperx>=3.1.0",
    "torch>=2.0.0",
    "googletrans==4.0.0rc1",
    "requests>=2.32.0",
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "pywebview>=4.4.0",
]
